Day Two!
Got the player anims working right. Still have an issue with the Controller bug.

The Controller Bug: Moving the player doesn't work right. When I try to move up or down, the animation is always left or right. It doesn't happen on mouse and keyboard tho.

That's an accessibility thing for me. I REALLY struggle with WASD with my fingers.  Especially when I'm high.